Add support for container runtime (#1232)
* Upgrade docker-client to latest

* Add support for runtime when creating a container

* Update user manual

* Add test for runtime host config

* Update runtime example in user guide

+++ helios-services/src/main/java/com/spotify/helios/agent/Supervisor.java
@@ -318,7 +318,7 @@ public class Supervisor {
       }
 
       // Check if the runner exited normally or threw an exception
-      final Result<Integer> result = runner.result();
+      final Result<Long> result = runner.result();
       if (!result.isSuccess()) {
         // Runner threw an exception, inspect it.
         final Throwable t = result.getException();
